fix severe memory leaks
The memory leaks happen when an attrib is uploaded to GLShader but is never freed. This can cause the program to eat up gigs of ram in seconds and freeze the kernel when rendering a high resolution sphere (and other primitives).

The bug can be reproduced by running `./clothsim -f ../scene/sphere.json -a 128 -o 128`.

Signed-off-by: Yuan Zhou <yvbbrjdr@berkeley.edu>